chore: consolidate pyarrow helpers
