+++ tests/unit/test__pandas_helpers.py
@@ -346,14 +346,14 @@ def test_bq_to_arrow_data_type_w_struct(module_under_test, bq_type):
         pyarrow.field("field04", pyarrow.int64()),
         pyarrow.field("field05", pyarrow.float64()),
         pyarrow.field("field06", pyarrow.float64()),
-        pyarrow.field("field07", module_under_test.pyarrow_numeric()),
-        pyarrow.field("field08", module_under_test.pyarrow_bignumeric()),
+        pyarrow.field("field07", _helpers.pyarrow_numeric()),
+        pyarrow.field("field08", _helpers.pyarrow_bignumeric()),
         pyarrow.field("field09", pyarrow.bool_()),
         pyarrow.field("field10", pyarrow.bool_()),
-        pyarrow.field("field11", module_under_test.pyarrow_timestamp()),
+        pyarrow.field("field11", _helpers.pyarrow_timestamp()),
         pyarrow.field("field12", pyarrow.date32()),
-        pyarrow.field("field13", module_under_test.pyarrow_time()),
-        pyarrow.field("field14", module_under_test.pyarrow_datetime()),
+        pyarrow.field("field13", _helpers.pyarrow_time()),
+        pyarrow.field("field14", _helpers.pyarrow_datetime()),
         pyarrow.field("field15", pyarrow.string()),
     )
     expected = pyarrow.struct(expected)
@@ -394,14 +394,14 @@ def test_bq_to_arrow_data_type_w_array_struct(module_under_test, bq_type):
         pyarrow.field("field04", pyarrow.int64()),
         pyarrow.field("field05", pyarrow.float64()),
         pyarrow.field("field06", pyarrow.float64()),
-        pyarrow.field("field07", module_under_test.pyarrow_numeric()),
-        pyarrow.field("field08", module_under_test.pyarrow_bignumeric()),
+        pyarrow.field("field07", _helpers.pyarrow_numeric()),
+        pyarrow.field("field08", _helpers.pyarrow_bignumeric()),
         pyarrow.field("field09", pyarrow.bool_()),
         pyarrow.field("field10", pyarrow.bool_()),
-        pyarrow.field("field11", module_under_test.pyarrow_timestamp()),
+        pyarrow.field("field11", _helpers.pyarrow_timestamp()),
         pyarrow.field("field12", pyarrow.date32()),
-        pyarrow.field("field13", module_under_test.pyarrow_time()),
-        pyarrow.field("field14", module_under_test.pyarrow_datetime()),
+        pyarrow.field("field13", _helpers.pyarrow_time()),
+        pyarrow.field("field14", _helpers.pyarrow_datetime()),
         pyarrow.field("field15", pyarrow.string()),
     )
     expected_value_type = pyarrow.struct(expected)
